fix delay when processing touchscreen events
BMC#14738, additional fix. Sometimes the screen got invoked couple of
seconds after blanking it even thought there were not any user activity.
This was caused by the code in event-input.c which suspended the event
reading for a while in certain situations after sending the activity
event. I removed the suspend code and instead use the input events time
as a filter so that only one activity event is send in maximum on each
second from the touchscreen events.

Signed-off-by: Mika Laitio <lamikr@pilppa.org>