Disable automatic suspend when MCE process exits
We must not leave the device in early suspend mode when the
MCE process exits.

Added an atexit handler for blocking suspend on normal exits.

Added SIGSEGV etc handlers that attempt to block suspend if
mce happens to crash.

Refactored signal handling to pass signals via pipe to be
processed outside the async signal handler context.

Added mce_abort() function that should be used instead of
_exit() and abort() in mce code.