Add cpu scaling governor policy
Makes it possible to configure mce so that "interactive" scaling
governor is used when device is running in normal UI state and
"performance" when device is booting up or shutting down.

Configuration via /etc/mce/NN-adaptation-name.ini files that specify
Default and Interactive sections, for example:
  [CPUScalingGovernorDefault]
  path1=/sys/devices/system/cpu/cpu*/cpufreq/scaling_governor
  data1=performance
  path2=
  [CPUScalingGovernorInteractive]
  path1=/sys/devices/system/cpu/cpu*/cpufreq/scaling_governor
  data1=interactive
  path2=

The mcetool utility can be used to override the policy via:
  mcetool -S automatic|performance|interactive
This setting persists over mce restarts / device reboots.

[mce] Add cpu scaling governor policy