Make it easier to configure LED patterns in multiple ini-files
Configuration entry "LEDPatterns" is renamed to "LEDPatternsRequired"
and is used for emitting diagnostic messages about patterns that should
be available but are not - instead of disabling patterns that are not
listed.

Separate config entry "LEDPatternsDisabled" is added for blocking
patterns even if they are configured.

Both "LEDPatternsRequired" and "LEDPatternsDisabled" lists can be
accumulated from multiple ini-files.

[mce] Make it easier to configure LED patterns in multiple ini-files