Add generic evdev input grab state machine
The idea is that state machine is told when grab/ungrab is desired,
but actual grab/ungrab happens when the input device is idle (finger
lifted from display, buttons released etc).

This should make it less likely that ui side that is also reading the
inputs misses touch / button release.