Grab touch input while display is off
Input is grabbed when display is powered off.

The grab is released when display is powered on and there are no
finger or palm touch detected.

This should stop the UI side from seeing touch events during
display power up and provides additional protection against
stray input if the device is held without covering the proximity
sensor.

[mce] Grab touch input while display is off