Ungrab volume keys when music playback is active
As there is no direct music playback is active information available,
heuristics based on volume limit audio policy decisions are used.

Based on that the volume keys are ungrabbed even if display is off or
ui is locked.

Also parsing of audio routing was changed, now all sink devices that
have name like "xxxforcall" are considered the same as "xxx". This
should make headset detection work the same whether "forcall" type
sinks are needed in the device or not.

[mce] Ungrab volume keys when music playback is active