Leave touch panel active during in-call proximity blanking
If touch panel is powered off, the touch detection does not work
reliably if there is finger/ear/chin on screen when the panel is
powered on again. Since that can break the touch blocking during
display power up in a situation where capacitive touch from chin
or ear is likely, the touch panel is not powered off during proximity
blanking even if the double tap detection is disabled.

[mce] Leave touch panel active during in-call proximity blanking