Start orientation sensor already when powering up the display
This way display power up and sensor ramp up happen in parallel and
we should already have up to date sensor data when the display state
transition is finished.

This needs to be revisited when LPM_ON and LPM_OFF display states are
taken into use to avoid unnecessary sensor starts and stops.