Support 100 brightness levels and up to 21 ALS profiles
Previously mce used brightness levels 1-5, that corresponded to 20, 40,
60, 80 or 100 percent brightness (manual mode) or selected one of the 5
ALS profiles available (automatic mode).

Now brightness level setting itself is percentage in manual mode and thus
allows user to have more freedom to select suitable brightness.

In automatic mode the percentage is mapped to one of the maximum of 21
ALS profiles. The profiles follow the same logic as the previous five
did, but having more of them allows user to select suitable compromise
between how low the brightness is allowed to go vs. how fast it should go
to the maximum brightness. There is also greater difference between the
profiles in "office lightning" conditions, which hopefully makes it more
apparent to users that the automatic mode is actually doing something.

Existing brightness settings in the 1-5 range are migrated on during mce
start up.

Also the dimmed display brightness is now dynamic and is always less
bright than the display on brightness.

[mce] Support 100 brightness levels and 21 ALS profiles. Fixes JB#13670