Move datapipes variables from mce.h to datapipe.c
Duplicates were generated in to every object file that was compiled
from source code that included mce.h. I guess this a) compiled
because the structures are not initialized b) worked because the
runtime linker resolved refs from the plugins to use variables
compiled into mce main binary.

Leave extern declarations to the header, but define the actual
variables only once in the datapipe module.

[mce] Move datapipes variables from mce.h to datapipe.c. Fixes JB#18917