Keep frame buffer device opened to keep display from powering off
When mce starts up, it will keep the fbdev open until actdead or
user mode is reached. This should keep the display powered on and
not lose content during bootup.

When shutdown is initiated (and not canceled), mce opens fbdev and
keeps it open for a while even after mce exits to allow shutdown
animation to stay visible.

[mce] Keep frame buffer device opened to keep display from powering off. Fixes JB#14341