Adjust display linger times after ending a call
When call ends the linger time specifies the length of time during
which the display will turn on / stay on if proximity is uncovered
before previous display state and normal blanking policy is restored.

For incoming calls the linger time should be long enough to allow the
user to select "call again" or similar continuation actions. For now
5 seconds is used.

For outgoing calls the linger time should be long enough to allow the
user to remove the device from ear. For now 2.5 seconds is used.

[mce] Adjust display linger times after ending a call. Fixes JB#20810