Use variable reaction time to uncovering of the proximity sensor
Previously 500ms delay was always used. While this makes stray touch
events less likely to occur when the phone is held at ear, it also
delays for example triggering of lpm display states when the phone
is removed from a pocket.

Now shorter delay is used unless there is an active call.