Add enablers for ui side dimming
When normal display backlight brightness is close to minimum allowed by
kernel/hw, it is not possible to make any visible difference between on
and dimmed state by further backlight brightness tuning.

Now mce sends a D-Bus signal that lipstick can use to make the graphics
rendered on screen darker when this happens.

[mce] Add enablers for ui side dimming. Fixes JB#21683