Check logging verbosity before evaluating format parameters
Vast majority of diagnostic logging from mce is not emitted to journal
on default verbosity. Spending cpu time evaluating format parameters
and pushing them to stack just to be discarded makes no sense.

Now verbosity level is checked 1st and potentially time consuming
parameter evaluation is done only if the message is going to be
emitted to the journal.

[mce] Check logging verbosity before evaluating format parameters. Fixes JB#22310