Do not release touch input blocking while proximity sensor is covered
If pressing power key or lipstick restart causes the display to turn
on while the device is in pocket, the unblank time touch detection might
not be enough from blocking touch events from reaching the ui and causing
pocket calls and similar undesirable actions.

Once the touch input has been grabbed, do not ungrab it while proximity
sensor is covered.

[mce] Do not release touch input blocking while proximity sensor is covered. Fixes JB#22587