Add graph describing transitions done at display state machine
The state machine logic dealing with display state changes can be
a bit hard to follow. I've maintained a dot-file that visualizes
the state machine operation - off the mce source tree. Having it
in the same place as the state machine implementation would make
things easier for everybody.

Add the dot-language graph to mce source tree.