Do not block on <--> dimmed display state transitions due to fading
Delaying transition to or from dimmed state makes the event eater
misbehave - events can get through instead of bouncing the display
back to on state or get ignored when they should not be.

If the transition is being made from states where ui is expected
to be active, do not delay the display state transition due to
ongoing brightness fade.