Improvements to sensor state handling when sensord is not available
If evdev source for als/ps is available amd sensord is not started, mce
ends up using sensor state that (possibly never powered up) sensors happen
to be reporting. This can break act dead mode totally if proximity sensor
default is covered (double tap does not work and alarms are not shown).

This fault was hidden for a while because sensord was erroneously started
in act dead too.

Add actual INITIAL state to relevant state machines instead of relying on
"sensord not available" IDLE states. This makes initial state transitions
to take place also in absence of sensord. This in turn is used to toggle
tracking_active flags for sensor reporting so that safe-guess-defaults
are used always when sensord is not alive regardless of whether we have
sensord data from sensord itself or from evdev sources.

[mce] Improvements to sensor state handling when sensord is not available. Fixes JB#24910