Make keypad slide policies work again
During tklock rewrite display blanking and autorelock policies based
on keypad slider state were broken.

Make keypad slide out/in work similarly as it used to with Nokia
N900 and N950 devices: Unblank display and hide lock screen when
keypad is slid out. If display was turned on due to keypad slide and
no touch input / calls are observed, blank display when the keypad
is slid in again.

[mce] Make keypad slide policies work again. Fixes JB#24391