Move dbus service tracking helpers from display plugin to mce-dbus
There are multitudes of local dbus service tracking variants scattered
around mce code base.

The lipstick state is needed by tklock core module and provided by
in theory optional display plugin.

Move the name owner tracking implemented in display plugin to mce-dbus
so that it can be used as common implementation for tracking all dbus
services that are essential for mce operation. Initially use it for
tracking availability of lipstick and compositor service.