Add plugin for tracking battery state via statefs
Performs similar task as existing battery_bme and battery_upower plugins,
but gets the battery data from files maintained by the statefs service.

Checks presense of required propery files in /run/state/namespaces/Battery
directory on mce startup. If files are not present, or become not available
later on due to stopping and/or restarting of statefs service, periodic
checks are made to detect when they come available again.

While the files available, the battery level, status and charging activity
state changes are tracked and pushed to datapipes in the form existing mce
logic is expecting.

[mce] Add plugin for tracking battery state via statefs. Contributes to JB#6273