Handle also transition to "unknown" battery state
The logic assumed that battery is in unknown state until statefs reports
transition to [charging, discharging, full, empty] is seen. However transitions
to unknown state are ignored and the previous state is kept.

Add "unknown" to set of expected Battery.State values.

Treat empty string as an alias for "unknown".

Unrecognized string values are treated similarly as "unknown", but a
warning is emitted to journal.