Increase the amount of configurable power key dbus actions to six
While refactoring the power key handler, it was assumed that two dbus
actions would be enough. Turns out that the assumption was wrong.


Increase the amount of customizable dbus actions from two to six, so
that it is now possible to configure one unique dbus action for each
combination of { short, long, double press } x {from display on, off }.

[mce] Increase the amount of configurable power key dbus actions to six. Fixes NEMO#781