Re-migrate brightness values after settings reset
For the brightness migration to work, it needs needs all three related
settings to be either in legacy or current state. Dealing with this
by using special logic during settings reset is efficient but not nice.

By forcing refreshing of all three values during sanitizing of the trio
we effectively do re-migration from reset defaults again. This adds
otherwise unneeded cpu load to normal brightness setting change, but
the amount of extra work is negligible.