Remove unused mce.ini configuration items
During the evolution of mce it has repeatedly happened that code for
some features has disappeared, but the related config items, macros
and in some cases even variables remain littering the source tree.

Remove all ini-file based configuration items, defines and variables
that are not actually used in the code.

[mce] Remove unused mce.ini configuration items. Fixes JB#25531