Using sensorfw for ambient light / proximity sensor control is not optional
The legacy sensor control methods are removed, only sensorfw can be used
for starting/stopping sensors. If sensord is not running on the device
or does not support some of the expected to be available sensors, mce
will still function and uses sensible fallback values for sensors.

Remove the build time configuration for disabling sensorfw usage.

[mce] Using sensorfw for ambient light / proximity sensor control is not optional. Fixes JB#27286