Add setting for autolock delay length
When the display blanks automatically via dimming, the autolock is
not applied immediately. During this so called grace period users can
get back to the app that was active before display was blanked.

While having this grace period is useful from user's point of view,
it also means applications do not get backgrounded and thus stay in
active state - which can cause issues. The hardcoded 30 second grace
period is a compromise between these conflicting needs.

Making the length of the grace period changeable at runtime allows users
to choose which of the two is more important for them.

Default stays at 30 seconds, but can be changed via:
  mcetool --set-autolock-delay=<seconds[.fraction]>

[mce] Add setting for autolock delay length. Fixes NEMO#787