Refactor proximity & autolock state machines
The proximity and delayed autolocking logic can get disabled due to
calls, alarms and other exceptional display on sequences. The code
is also hard to follow due to use of triggering based on changes to
individual states.

Refactor the code so that there is one function for controlling
whether the state machine should be active and one function for
evaluating whether the conditions for triggering tklocking apply.

[mce] Refactor proximity & autolock state machines. Contributes to JB#21182