Cap automatic blanking timeouts in act dead mode
The display blanking settings are used in the act dead mode too.
If user has configured long blanking timeouts, this can adversely
affect charging in the act dead mode.

Apply upper limit of 15 seconds for dimming timeouts and 3 seconds for
blanking timeouts while the device is in the act dead mode.

[mce] Cap automatic blanking timeouts in act dead mode. Fixes JB#28709