Wakeup from suspend to end autolock grace period
When display is first dimmed and then blanked based on user inactivity,
the tklock is not immediately set. This so called grace period allows
user some time wake up the display and still get back to the application
that was active before blanking. But in case the device suspends, it
can also mean that things like device lock and application activity
states evaluation gets delayed until the device wakes up from suspend
to turn on the display - which can cause for example visual hiccups
and missing notification leds.

Use the mce-hbtimer functionality to force wakeup from suspend when
the grace period is up.

[mce] Wakeup from suspend to end autolock grace period. Fixes JB#28706