Provide separate datapipes for raw and filtered ALS data
By default the ALS data is median filtered, which makes it too slow
reacting to be used for things like substitute for proximity sensor
and only the filtered state can be subscribed to.

Make it possible for mce logic to subscribe to also raw als data
by providing a separate data pipe for it.

[mce] Provide separate datapipes for raw and filtered ALS data. Contributes to JB#29011