Use ambient light to avoid false positives with magnetic cover sensors
Depending on shielding and magnet strength the magnetic cover sensors
can get triggered by a magnet underneath the device too. This is
problematic since the policy for the lid sensor is not to allow
turning display on for any reason - power key, double tap etc will
not function.

Ignore cover closed events while display is on unless they happen in
close proximity to ambient light level dropping close to zero.

Opening the cover is not subject to ALS rules and always turns on
the display.

[mce] Use ambient light to avoid false positives with magnetic cover sensors. Contributes to JB#29011