Make lid sensor policy work similarly to proximity sensor
The lid sensor policy is implemented as a display state filter that
does not allow turning on the display for any reason while the
sensor is in closed position. This can be problematic in case there
are false positives from the lid sensor for any reason.

Instead of global display state filtering apply similar case by
case rules when the lid sensor is in closed position as what are
used for the proximity sensor:
* Display power up requests via D-Bus are ignored
* Calls/alarms/notifications will not turn on display
* By default power key can be used to turn display on
* Sliding keyboard out does not turn display on
* Touch input can be optionally blocked
* LPM display state is not allowed

The configurable behavior follows the settings for the proximity sensor.

[mce] Make lid sensor policy work similarly to proximity sensor. Fixes JB#29011