Reset lid-sensor-is-working status on triple powerkey press
Similarly to proximity sensor, it is now possible to configure also
the lid sensor to block power key events. This makes it difficult to
recover from situations where the sensor suddenly starts misbehaving.

Triple powerkey press can already be used to force cached proximity
sensor state to not-covered position. Use the same triggering mechanism
to reset lid-sensor-is-working status. Effectively the sensor is ignored
until the next time it switches to open position.

[mce] Reset lid-sensor-is-working status on triple powerkey press. Contributes to JB#29011