Reset als history after enabling the sensor
The intent is to reset als history to fresh data when it comes available
from the sensor. However the sensorfw module feeds the last known value
when the sensor is enabled and this ends up being used for history buffer
initialization and slows the reaction to actual change in lightning
condition in case it has changed during sensor down time.

Conditionally reset the history buffer after enabling the sensor. This
way the buffer gets reset using the next value that differs from the
previously seen one, median filtering starts off from the freshly
obtained value and display should wake up to correct brightness.

[mce] Reset als history after enabling the sensor. Contributes to JB#30090