[display] Add setting for blanking pause mode. Fixes JB#27486
Usually navigation applications use blank prevent to keep display on.
Sometimes - say when navigating while driving with bike - it might
be desirable to still allow display to dim (or blank) to save battery.

Add setting for controlling blank prevent handling to make it possible
to allow display to dim or even blank while application is using blank
prevention D-Bus interface.

The supported modes are:

  * keep-on: blank prevent sessions keep the display on as before
  * allow-dim: display can be dimmed even if blank prevent is used
  * disabled: display can be blanked even if blank prevent is used

By default the "keep-on" mode is used and blank prevention works just
like it did before the changes. The mode can be changed via:

  mcetool --set-blank-prevent-mode=<disabled|keep-on|allow-dim>

The value of the setting persists over mce / device restarts.