[doubletap] Disable double tap detection when lid is closed. Contributes to JB#27807
If the lid sensor indicates that the whole display is covered, it would
make sense to disable double tap detection to save some power.

Utilize the already existing logic and configuration that disables
double tap detection based on proximity sensor to perform similar
function for the based on filtered lid sensor state too.