[proximity] Optionally make proximity sensor behave like lid sensor. Fixes MER#1201
On devices that have a lid cover sensor, mce uses it for turning the
display on/off when the cover is opened/closed. On devices that do not
have a dedicated lid sensor, users might prefer using proximity sensor
for that purpose even if it has some side effects.

Add a setting for making mce treat proximity sensor act as if were
a lid sensor.

The setting is disabled by default and can be changed with:

  mcetool --set-ps-acts-as-lid=<enabled|disabled>

The value of the setting persists over mce / device restarts.