[tklock] Add setting for lid open/close actions. Fixes MER#1202
The actions taken when lid sensor indicates that cover is closed or opened
are hard coded without any possibility for users to tweak the behavior.

Add settings for tweaking of actions to taken when cover is closed:
  * disabled - do nothing
  * blank    - just turn off the display
  * tklock   - turn off display and activate lockscreen

Default is 'tklock' and can be changed via:
  mcetool --set-lid-close-actions=<disabled|blank|tklock>


Add settings for tweaking of actions to taken when cover is opened:
  * disabled - do nothing
  * unblank  - just turn on the display
  * tkunlock - turn on display and deactivate lockscreen (if possible)

Default is 'unblank' and can be changed via:
  mcetool --set-lid-open-actions=<disabled|unblank|tkunlock>

The values of the settings persists over mce / device restarts.