[activity] Allow both monotonic and realtime input event time stamps. Fixes JB#31540
For the purposes of rate-limiting user activity renew actions mce checks
that seconds in input event time stamp grow. This works when input
devices are using CLOCK_MONOTONIC as time base, but breaks user activity
monitoring if CLOCK_REALTIME is used and system time is moved backwards.

As any change in the seconds value in the time change is enough for rate
limiting to function, use inequality check instead of than less than check.