[tklock] Add setting to disable input device grabbing. Fixes MER#1250
There have been issues with unresponsive touch screen in various porting
projects. It is difficult to tell whether problems occur because

a) mce is denying touch events from reaching ui, or
b) ui side does not process touch events as expected

To ease debugging of such issues, add a setting that can be used to
control whether mce is allowed to grab input devices or not.

The setting is enabled by default and can be changed with:

  mcetool --set-input-policy-mode=<enabled|disabled>

The value of the setting persists over mce / device restarts.