[dsme] Move dsme signal handling to mce-dsme module. Fixes JB#28901
Due to historical reasons pre-shutdown signals dsme is sending are
handled in display plugin. This is problematic because there is a
dependency between system and shutdown-in-progress states which now
are handled in multiple ways in more than one place.

Move signal handlers to mce-dsme module and unify handling of
shutdown-in-progress state.