[tklock] Show device unlock view if tklock removal must be denied. Fixes MER#1324
Previously removal of tklock while device lock is active meant that the
device unlock view is shown to the user. Now, since the device unlocking
happens in the context of lockscreen (i.e. tklock is set), removal of
tklock can't be allowed while device lock is active. This means that
the behavior tweaks (double tap, lid open, power key) allowing bypassing
of the lockscreen no longer work.

If the device is configured as having devicelock-in-lockscreen, issue
show-device-unlock-view ipc when removal of tklock is requested while
device lock is active.