[tklock] Add settings for tuning keyboard slide open/close policy. Contibutes to MER#1321
Many users wish to tweak what happens when sliding keyboard is opened
or closed, but reaction to keyboard slide events is hard coded in mce.

Add settings for controlling when and how mce should react to keyboard
slide open/close events.

The default behavior is not changed, mce still does on slide
  * open: unblank display - if proximity sensor is not covered
  * close: blank and activate lockscreen - if display was unblanked
    due to keyboard slide open

The defaults can be changed by installing configuration files to
/etc/mce directory or via mcetool options:
  --set-kbd-slide-open-trigger=<never|no-proximity|always>
  --set-kbd-slide-open-actions=<disabled|unblank|tkunlock>
  --set-kbd-slide-close-trigger=<never|after-open|always>
  --set-kbd-slide-close-actions=<disabled|blank|tklock>

The values of the settings persists over mce / device restarts.

Also make sure closing keyboard slide does not blank the display when
there are active alarms or calls as it might interfere with display
state handling in a situation where power key can't be used to bring
the display back on.