[display] Add setting for kbd slide inhibit mode. Contributes to MER#1321
There are users that wish that the display would not blank when keyboard
slide is in open state.

Add inhibit mode setting that can be used to disable dimming / blanking
blanking timers when kbd slide is in open position.

By default kbd slide status does not affect display blanking timers.

The default can be changed by installing configuration files to
/etc/mce directory or via mcetool option:
  --set-kbd-slide-inhibit-mode=<disabled|stay-on-when-open|
                                stay-dim-when-open>

The value of the setting persists over mce / device restarts.