[display] Refactor adaptive dimming handling. Contributes to MER#1321
The adaptive dimming delay is implemented by maintaining indices to
allowed-dim-timeouts list. This implicitly forces it so that it is not
possible to select dim timeout outside the listed values.

Track last used dim-timeout and do adaptive adjustment based on the
value instead of index. This allows selection of dim-timeouts that
are in between/beyond the "allowed" ones.

Also fix inactivity delays so that bootup, system state and adaptive
dimming special cases are taken into account.