[display] Add setting for dim timeout with hw keyboard available. Fixes MER#1321
There are users that wish that the display would stay on longer when
a hw keyboard is attached / slider keyboard is open.

Add separate dim-timeout setting that is used when keyboard is available.
Setting it to zero value means: Use the regular dim-timeout even if a hw
keyboard is available.

By default the value is set to zero. This can be changed by installing
configuration files to /etc/mce directory or with mcetool option:
  --set-dim-with-kbdtimeout=<secs>

The value of the setting persists over mce / device restarts.