[tklock] Add setting for disabling unblank to lockscreen animation. Contributes to JB#30849
A lot of users think the animated unblank to lockscreen is too slow.

While there is at present no way to directly control the animation that
happens on the ui side, it does get implicitly disabled when display is
unblanked due to alarms/calls/notifications. Utilize this and optionally
fake notification unblank when a double tap or power key press is used
to wake up the device.

Add setting for enabling/disabling lockscreen unblank animations.

By default the animations are enabled, but this can be changed via
mcetool option:
  --set-lockscreen-animation=<enabled|disabled>

The value of the setting persists over mce / device restarts.