[powerkey] Do not bypass lockscreen if proximity sensor is covered. Fixes JB#32668
By default powerkey presses are acted on regardless of the proximity
sensor state. This can cause problems especially if powerkey is
configured to bypass lockscreen.

Do not allow "tkunlock" powerkey action to be taken while proximity
sensor is in covered state. This way users can alway unblank the
display even if the proximity sensor is covered, but touch interaction
is needed to get access to applications.