[dbus] Move dbus_any_t union declaration to mce-dbus.h
Essentially the same declaration is made in several source files and
is used to simplify D-Bus message construction and parsing.

Move the declaration to mce-dbus.h header file and remove redundant
declarations from source files.

Rename the union declared in the radiostates plugin as it is used in
slightly different manner and needs to have non-const pointers.