[tklock] Disable state restore on start of noanim blanking policy. Fixes MER#1399
The UIEXC_NOANIM blanking policy disables display state restore if it gets
to the top of the priority queue before it gets automatically disabled by
transition to the next stable display state. Now, Because the LPM_ON -> ON
transition is immediate and does not need to go through the transitional
POWER_UP state, the disabling of the display state restore does not happen
when noanim is invoked while lpm ui is visible.

Disable state restore immediately if there are no other exceptions active
when UIEXC_NOANIM is invoked.