[led] Do not log pattern details when evaluating active pattern. Fixes MER#1412
While the logging is useful when explicitly debugging the led plugin,
this particular line generates so much noise that journald starts
skipping log entries from mce and reduces the usability of the verbose
logging mode.

Leave logging in code a in form that is easy to activate when doing manual
builds, but do not compile it in by default.