[touch] Add handler for multitouch protocol A and B. JB#33644
The existing touch screen input event handling in mce is implemented in
several places and in general does handle differences between A and B
protocols in robust manner - and assumes that there will be only one
touch/mouse input device available.

Implement logic for tracking touch input state / double tap detection in
a manner that allows the code to be re-used for several input devices
when/if needed.