[input] Add setting to allow/deny input device grabbing. JB#33644
Grabbing touch input events is problematic - especially with devices
that use stateful protocol B for reporting. In general it would be better
to allow lipstick to see all the touch events but ignore them while
display is not in fully powered up state.

Add a setting that can be used to instruct mce not to grab touch input
device nodes. By default the grabbing is allowed.

Normally platform specific configuration files are expected to be used
when needed, but users can also tweak the setting via mcetool option
--set-input-grab-allowed.