[leds] Disable DisplayDimmed led pattern by default. Fixes JB#34543
Before compositor side dimming was implemented, it was very hard to
spot display dimming while the device was in low light conditions.
To avoid display blanking without any notice, a pale blue led color
was used as about-to-blank-screen indicator.

With compositor blending graphics on screen towards black, the
display dimming is always visually detectable and we can disable
the DisplayDimmed led pattern as it seems to be more confusing
rather than helpful.

Anyone who still would like to have the led effect, can use
  mcetool --enable-led-pattern=PatternDisplayDimmed
to restore the old behavior.